Refactorando os métodos das classes

+++ src/sgbf/controlo/ConAcervo.java
@@ -9,6 +9,8 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import sgbf.modelo.ModAcervo;
+import sgbf.modelo.ModAcervoFisico;
 import sgbf.modelo.ModEstante;
 import sgbf.util.UtilControloExcessao;
 import sgbf.util.UtilIconesDaJOPtionPane;
@@ -21,18 +23,29 @@ public class ConAcervo extends ConCRUD {
     
     @Override
     public boolean registar(Object objecto_registar, String operacao) {
-        //ModEstante estanteMod = (ModEstante)objecto_registar;
+        ModAcervo acervoMod = (ModAcervoFisico)objecto_registar;
         try{
             super.query = "INSERT INTO tcc.acervos (titulo, subtittulo, tipo_acervo, formato, edicao, volume,"
                         + " numero_paginas, codigo_barra, isbn, idioma, ano_lancamento, sinopse, endereco_acervo,"
                         + " categoria_idcategoria, Editora_idEditora, Estoque_idEstoque)"
                         + " VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
-            /*super.preparedStatement = super.caminhoDaBaseDados.baseDeDados(operacao).prepareStatement(query);
-            super.preparedStatement.setString(1, estanteMod.getDesignacao());
-            super.preparedStatement.setString(2, estanteMod.getDescricao());
-            super.preparedStatement.setByte(3, estanteMod.getLinha());
-            super.preparedStatement.setByte(4, estanteMod.getColuna());
-            super.preparedStatement.setInt(5, estanteMod.getAreaMod().getIdArea());*/
+            super.preparedStatement = super.caminhoDaBaseDados.baseDeDados(operacao).prepareStatement(query);
+            super.preparedStatement.setString(1, acervoMod.getTitulo());
+            super.preparedStatement.setString(2, acervoMod.getSub_titulo());
+            super.preparedStatement.setString(3, acervoMod.getTipo_acervo());
+            super.preparedStatement.setString(4, acervoMod.getFormato());
+            super.preparedStatement.setByte(5, acervoMod.getEdicao());
+            super.preparedStatement.setByte(6, acervoMod.getVolume());
+            super.preparedStatement.setInt(7, acervoMod.getNumero_paginas());
+            //super.preparedStatement.setByte(8, acervoMod.get());
+            //super.preparedStatement.setByte(9, acervoMod.getEdicao());
+            super.preparedStatement.setString(10, acervoMod.getIdioma());
+            super.preparedStatement.setInt(11, acervoMod.getAno_lancamento());
+            //super.preparedStatement.setByte(12, acervoMod.getEdicao());
+            //super.preparedStatement.setByte(13, acervoMod.getEdicao());
+            super.preparedStatement.setInt(14, acervoMod.getCategoriaMod().getIdCategoria());
+            //super.preparedStatement.setByte(15, acervoMod.get());
+            super.preparedStatement.setInt(16, acervoMod.getEstoqueMod().getIdEstoque());
             return !super.preparedStatement.execute();
         }catch(SQLException erro){
             throw new UtilControloExcessao("Erro ao "+operacao+" Acervo !\nErro: "+erro.getMessage(), operacao,UtilIconesDaJOPtionPane.Erro.nomeDaImagem());
@@ -43,18 +56,29 @@ public class ConAcervo extends ConCRUD {
 
     @Override
     public boolean alterar(Object objecto_alterar, String operacao) {
-        //ModEstante estanteMod = (ModEstante)objecto_alterar;
+        ModAcervo acervoMod = (ModAcervoFisico)objecto_alterar;
         try{
             super.query = "UPDATE tcc.acervos set titulo=?, subtittulo=?, tipo_acervo=?, formato=?, edicao=?, volume=?,"
                         + " numero_paginas=?, codigo_barra=?, isbn=?, idioma=?, ano_lancamento=?, sinopse=?, endereco_acervo=?,"
                         + " categoria_idcategoria=?, Editora_idEditora=?, Estoque_idEstoque=? where idAcervos=?";
-            /*super.preparedStatement = super.caminhoDaBaseDados.baseDeDados(operacao).prepareStatement(query);
-            super.preparedStatement.setString(1, estanteMod.getDesignacao());
-            super.preparedStatement.setString(2, estanteMod.getDescricao());
-            super.preparedStatement.setByte(3, estanteMod.getLinha());
-            super.preparedStatement.setByte(4, estanteMod.getColuna());
-            super.preparedStatement.setInt(5, estanteMod.getAreaMod().getIdArea());
-            super.preparedStatement.setInt(6, estanteMod.getIdEstante());*/
+            super.preparedStatement = super.caminhoDaBaseDados.baseDeDados(operacao).prepareStatement(query);
+            super.preparedStatement.setString(1, acervoMod.getTitulo());
+            super.preparedStatement.setString(2, acervoMod.getSub_titulo());
+            super.preparedStatement.setString(3, acervoMod.getTipo_acervo());
+            super.preparedStatement.setString(4, acervoMod.getFormato());
+            super.preparedStatement.setByte(5, acervoMod.getEdicao());
+            super.preparedStatement.setByte(6, acervoMod.getVolume());
+            super.preparedStatement.setInt(7, acervoMod.getNumero_paginas());
+            //super.preparedStatement.setByte(8, acervoMod.get());
+            //super.preparedStatement.setByte(9, acervoMod.getEdicao());
+            super.preparedStatement.setString(10, acervoMod.getIdioma());
+            super.preparedStatement.setInt(11, acervoMod.getAno_lancamento());
+            //super.preparedStatement.setByte(12, acervoMod.getEdicao());
+            //super.preparedStatement.setByte(13, acervoMod.getEdicao());
+            super.preparedStatement.setInt(14, acervoMod.getCategoriaMod().getIdCategoria());
+            //super.preparedStatement.setByte(15, acervoMod.get());
+            super.preparedStatement.setInt(16, acervoMod.getEstoqueMod().getIdEstoque());
+            super.preparedStatement.setInt(17, acervoMod.getIdAcervo());
             return !super.preparedStatement.execute();
         }catch(SQLException erro){
             throw new UtilControloExcessao("Erro ao "+operacao+" Acervo !\nErro: "+erro.getMessage(), operacao,UtilIconesDaJOPtionPane.Erro.nomeDaImagem());
@@ -65,23 +89,34 @@ public class ConAcervo extends ConCRUD {
 
     @Override
     public boolean remover(Object objecto_remover, String operacao) {
-        //ModEstante estanteMod = (ModEstante)objecto_remover;
+         ModAcervo acervoMod = (ModAcervoFisico)objecto_remover;
         try{
-            super.query = "delete from tcc.acervos where idAcervos=?";
-            /* super.preparedStatement = super.caminhoDaBaseDados.baseDeDados(operacao).prepareStatement(query);
-            super.preparedStatement.setInt(1,estanteMod.getIdEstante());*/
-            return !super.preparedStatement.execute();
+            if(this.temDadosRelacionados(acervoMod, operacao)){
+               throw new UtilControloExcessao("Esta operacao não pode ser executada\nO Acervo seleccionado possui registo ! ", operacao, UtilIconesDaJOPtionPane.Erro.nomeDaImagem());
+            }else{
+                super.query = "delete from tcc.acervos where idAcervos=?";
+                super.preparedStatement = super.caminhoDaBaseDados.baseDeDados(operacao).prepareStatement(query);
+                super.preparedStatement.setInt(1,acervoMod.getIdAcervo());
+                return !super.preparedStatement.execute();
+            }
         }catch(SQLException erro){
            throw new UtilControloExcessao("Erro ao "+operacao+" Acervo !\nErro: "+erro.getMessage(), operacao, UtilIconesDaJOPtionPane.Erro.nomeDaImagem());
         }finally{
             super.caminhoDaBaseDados.fecharTodasConexoes(preparedStatement, setResultset, operacao);
         }
     }
+    
+    private boolean temDadosRelacionados(ModAcervo acervoMod, String operacao) throws SQLException{
+        super.query = "select *from itensSolicitados where Acervos_idAcervos=?";
+        super.preparedStatement = super.caminhoDaBaseDados.baseDeDados(operacao).prepareStatement(super.query);
+        super.preparedStatement.setInt(1, acervoMod.getIdAcervo());
+        return super.setResultset.next();
+    }
 
     @Override
     public List<Object> listarTodos(String operacao) {
         List<Object> todosRegistos = new ArrayList<>();
-        try{
+        /*try{
             super.query = "select * from tcc.acervos order by titulo, data_modificacao asc";
             super.preparedStatement = super.caminhoDaBaseDados.baseDeDados(operacao).prepareStatement(query);
             super.setResultset = super.preparedStatement.executeQuery();
@@ -93,13 +128,14 @@ public class ConAcervo extends ConCRUD {
             throw new UtilControloExcessao("Erro ao "+operacao+" Acervo(s) !\nErro: "+erro.getMessage(), operacao, UtilIconesDaJOPtionPane.Erro.nomeDaImagem());
         }finally{
             super.caminhoDaBaseDados.fecharTodasConexoes(preparedStatement, setResultset, operacao);
-        }
+        }*/
+        return todosRegistos;
     }
 
     @Override
     public List<Object> pesquisar(Object objecto_pesquisar, String operacao) {
         List<Object> todosRegistosEncontrados = new ArrayList<>();
-        ModEstante estanteMod = (ModEstante)objecto_pesquisar;
+        /*ModEstante estanteMod = (ModEstante)objecto_pesquisar;
         try{
             super.query = "select * from tcc.acervos where idAcervos=? or "
                         + "titulo like '%"+estanteMod.getDesignacao()+"%'";
@@ -114,7 +150,8 @@ public class ConAcervo extends ConCRUD {
             throw new UtilControloExcessao("Erro ao "+operacao+" Acervo(s) !\nErro: "+erro.getMessage(), operacao,UtilIconesDaJOPtionPane.Erro.nomeDaImagem());
         }finally{
             super.caminhoDaBaseDados.fecharTodasConexoes(preparedStatement, setResultset, operacao);
-        }
+        }*/
+        return todosRegistosEncontrados;
     }
     
     private Object pegarRegistos(ResultSet setResultset,String operacao) throws SQLException{
